🦄 refactor:Ajout de la fonction de suppression de sprite pour les personnages en ciblant correctement le champ "delete_sprite" associé au formulaire du bon personnage par "character_id".
This commit refactors the character form in the SceneCreation view to allow users to delete a character's sprite. If the user confirms the deletion, the corresponding character's sprite is set to an empty string. This change improves the user experience by providing an option to remove a character's sprite easily.